requirements.txt
google-generativeai==0.3.2
pytest==7.3.1
python-dotenv==1.0.0
pandas==1.5.3
black==23.3.0
mypy==1.3.0
click==8.1.3